- For API consistency made split size an int instead of a long

+++ src/main/com/mongodb/hadoop/MongoConfig.java
@@ -265,11 +265,11 @@ public class MongoConfig {
         MongoConfigUtil.setSkip(_conf, skip);
     }
 
-    public long getSplitSize() {
+    public int getSplitSize() {
         return MongoConfigUtil.getSplitSize(_conf);
     }
 
-    public void setSplitSize(long value) {
+    public void setSplitSize(int value) {
         MongoConfigUtil.setSplitSize(_conf, value);
     }
